1. Added Enemy Bullet shoot and collision 2. Fixed minor issues. 3. GAME COMPLETE!!

+++ app/src/main/java/ca/georgebrown/game2011/arcadegame/Activities/GamePlayActivity.java
@@ -2,9 +2,11 @@ package ca.georgebrown.game2011.arcadegame.Activities;
 
 
 import android.app.Activity;
+import android.graphics.Typeface;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.LinearLayout;
+import android.widget.TextView;
 
 import ca.georgebrown.game2011.arcadegame.GameView.GameCanvasView;
 import ca.georgebrown.game2011.arcadegame.R;
@@ -17,15 +19,22 @@ public class GamePlayActivity extends Activity {
 
     GameCanvasView gameCanvas;
     LinearLayout pauseMenuView;
+    TextView gamePauseTextView;
+
+    boolean didPlayerExitGame = false;
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_play);
         gameCanvas = findViewById(R.id.gameCanvas);
         pauseMenuView = findViewById(R.id.pause_menu_view);
-        pauseMenuView.setVisibility(View.INVISIBLE);
+        gamePauseTextView = findViewById(R.id.game_pause_text);
 
+        pauseMenuView.setVisibility(View.INVISIBLE);
         gameCanvas.pauseMenuView = this.pauseMenuView;
+
+        Typeface tf = Typeface.createFromAsset(getAssets(), "fonts/anabelle_script_light.otf");
+        gamePauseTextView.setTypeface(tf);
 //        gameCanvas.resume();
     }
 
@@ -35,6 +44,7 @@ public class GamePlayActivity extends Activity {
     }
 
     public void onExitGameButtonClicked(View view){
+        didPlayerExitGame = true;
         finish();
     }
 
@@ -53,6 +63,8 @@ public class GamePlayActivity extends Activity {
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        android.os.Process.killProcess(android.os.Process.myPid());
+        if (didPlayerExitGame){
+            android.os.Process.killProcess(android.os.Process.myPid());
+        }
     }
 }
